Separated additional terms (LogicList and Symbol) and test utils to different module
