+++ klogic-core/src/main/kotlin/org/klogic/core/Constraint.kt
