+++ klogic-core/src/main/kotlin/org/klogic/core/Goal.kt
