+++ klogic-core/src/main/kotlin/org/klogic/core/State.kt
