+++ klogic-core/src/main/kotlin/org/klogic/core/Stream.kt
