+++ klogic-core/src/main/kotlin/org/klogic/core/Substitution.kt
