+++ klogic-core/src/main/kotlin/org/klogic/core/Term.kt
