+++ klogic-core/src/test/kotlin/org/klogic/core/streams/BindTest.kt
