+++ klogic-core/src/test/kotlin/org/klogic/core/streams/MplusTest.kt
