+++ klogic-utils/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("org.klogic.klogic-base")
+    `java-test-fixtures`
+}
+
+dependencies {
+    implementation(project(":klogic-core"))
+    testFixturesImplementation(project(":klogic-core"))
+}
