--- src/test/kotlin/org/klogic/core/unify/WalkTest.kt
@@ -1,19 +1,0 @@
-package org.klogic.core.unify
-
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Test
-import org.klogic.core.Substitution.Companion.of
-import org.klogic.terms.Nil.nilLogicList
-import org.klogic.terms.Symbol
-import org.klogic.utils.x
-
-class WalkTest {
-    @Test
-    fun testWalk() {
-        val substitution = of(x to nilLogicList<Symbol>())
-        val term = x.walk(substitution)
-
-        val expected = nilLogicList<Symbol>()
-        assertEquals(expected, term)
-    }
-}
